server/requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.2
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
uritemplate==3.0.0
urllib3==1.21.1
uWSGI==2.0.15
django-rest-auth==0.9.1
django-rest-auth[with_social]
django-filter==1.0.4